backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
openai==1.12.0
httpx==0.27.0
python-dotenv==1.0.0
